feat(tfm-reporting): Code abstraction
